% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_net_sentiment.R
\name{calc_net_sentiment}
\alias{calc_net_sentiment}
\title{Calculate Net Sentiment}
\usage{
calc_net_sentiment(processed_comments, comments)
}
\arguments{
\item{processed_comments}{data frame, result of \code{process_comments}}

\item{comments}{data frame, result of 'extract_comments'}
}
\value{
data frame
}
\description{
Calculate Net Sentiment
}
