% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_comments.R
\name{extract_comments}
\alias{extract_comments}
\title{Extract Comments}
\usage{
extract_comments(urls)
}
\arguments{
\item{urls}{urls of reddit posts}
}
\value{
data frame with comments
}
\description{
Extract Comments
}
