% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comments.R
\name{plot_comments}
\alias{plot_comments}
\title{Plot Top Words in Processed Comments}
\usage{
plot_comments(comments, type)
}
\arguments{
\item{comments}{data frame, result of \code{process_comments}}

\item{type}{chr string, one of: "col", "wordcloud"}
}
\value{
ggplot object
}
\description{
Plot Top Words in Processed Comments
}
