% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sentiment.R
\name{plot_sentiments}
\alias{plot_sentiments}
\title{Plot Sentiments}
\usage{
plot_sentiments(sentiments)
}
\arguments{
\item{sentiments}{data frame, result of \code{calc_net_sentiment}}
}
\value{
ggplot object
}
\description{
Plot Sentiments
}
