% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_comments.R
\name{process_comments}
\alias{process_comments}
\title{Process Comments for tidytext}
\usage{
process_comments(comments, other_stop_words = "gt")
}
\arguments{
\item{comments}{data frame, result of \code{extract_comments} function}

\item{other_stop_words}{character vector of words to be removed, eg. for Reddit its 'gt', an artifact from inline HTML in the comments, for multiple use c("word1", "word2")}
}
\value{
data frame
}
\description{
Process Comments for tidytext
}
