% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sentiment.R
\name{run_sentiment}
\alias{run_sentiment}
\title{Run Sentiment Analysis}
\usage{
run_sentiment(
  search,
  max_results = 50,
  sort_by = "relevance",
  posts_from = "all"
)
}
\arguments{
\item{search}{chr string}

\item{max_results}{numeric}

\item{sort_by}{chr string}

\item{posts_from}{chr string}
}
\value{
list of data frames
}
\description{
Run Sentiment Analysis
}
