% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_reddit.R
\name{search_reddit}
\alias{search_reddit}
\title{Search Reddit}
\usage{
search_reddit(
  search,
  max_results = 50,
  sort_by = "relevance",
  posts_from = "all"
)
}
\arguments{
\item{search}{character string, search query passed to reddit}

\item{max_results}{numeric, maximum number of reddit results, max is 100}

\item{sort_by}{character string, one of reddit's sorting algorithms e.g 'new', 'comments', 'relevance', 'hot', 'top'}

\item{posts_from}{character string, filter post based on when they were posted, one of: "hour", "day", "week", "month", "year", "all"}
}
\value{
data frame with post title, url, number of comments
}
\description{
Search Reddit
}
